+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Fixed
 - Passing start date when checking availability no longer makes that date the primary rules start date.
 
+## Added
+- Added `Event::getNextAvailableSlot()`
+
 ## 1.0.0-alpha.4
 ### Fixed
 - Fixed availability error when database has prefix